requirements_dev.txt
-r requirements.txt

# You already have these
Django>=1.6.0
djangorestframework>=2.3,<2.3.99

# These are just used for development

mock
wsgiref>=0.1.2
selenium
requests

# This is used for generating documentation suitable for pypi
pypandoc
